Rename SignUpAlumno file
